refactor: move resume helpers into SDK-owned lib module
Follows the shape of #3712/#3713 (polling helpers in SDK-owned files):
the byte-range resume helpers, state object, and body readers now live in
the handwritten src/openai/lib/_resumable.py, leaving only the retry-loop
integration (~120 lines) in the template-owned _base_client.py. No
behavior change; the tests are unchanged apart from imports.
